Midterm/requirements.txt
# Core ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Database
snowflake-connector-python>=3.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

langchain-openai

streamlit
plotly
numpy
langragph
fastapi
uvicorn[standard]
pydantic
pandas
